requirements.txt
apex==0.9.10dev
laplace==0.1
laplace_torch==0.1a2
netcal==1.2.1
networks==0.3.7
numpy==1.20.2
ood_metrics==0.2.9
pandas==1.2.2
Pillow==9.2.0
pytorch_lightning==1.5.10
scikit_learn==1.1.1
torch==1.10.1
torchlars==0.1.2
torchvision==0.11.2
